Don't accidentally overwrite a file that exists

+++ init.ps1
@@ -11,10 +11,12 @@ if (-not (Test-Path -Path ".\puzzles")) {
     New-Item -Path ".\puzzles" -ItemType "directory"
 }
 
-$contents = Get-Content -Path ".\boilerplate\solve.txt"
-
 $alphanum = $puzzle.PadLeft(4-$puzzle.Length, "0")
-$contents = $contents -replace "#0#", $puzzle
-$contents = $contents -replace "#1#", $alphanum
+if (-not (Test-Path -Path ".\puzzles\$alphanum.go")) {
 
-Set-Content -Path ".\puzzles\$alphanum.go" -Value $contents
+    $contents = Get-Content -Path ".\boilerplate\solve.txt"
+    $contents = $contents -replace "#0#", $puzzle
+    $contents = $contents -replace "#1#", $alphanum
+
+    Set-Content -Path ".\puzzles\$alphanum.go" -Value $contents
+}
